api/views.py
from django.shortcuts import render
from rest_framework import viewsets

from .models import *
from .serializers import *

class ProjectViewSet(viewsets.ModelViewSet):

  queryset = Project.objects.all()
  serializer_class = ProjectSerializer
