requirements.txt
boto==2.48.0
dj-database-url==0.4.2
Django==2.0.2
django-cors-headers==2.1.0
django-storages==1.6.5
djangorestframework==3.7.7
psycopg2-binary==2.7.4
pytz==2018.3
whitenoise==3.3.1
